+++ pytest.ini
@@ -5,5 +5,3 @@ log_cli=true
 log_level=DEBUG  
 log_format = %(asctime)s %(levelname)s %(message)s  
 log_date_format = %Y-%m-%d %H:%M:%S
-testpaths =
-    audio/tests
