+++ save_as_docx.py
@@ -1,0 +1,12 @@
+from docx import Document
+
+def save_as_docx(minutes, filename):
+    doc = Document()
+    for key, value in minutes.items():
+        # Replace underscores with spaces and capitalize each word for the heading
+        heading = ' '.join(word.capitalize() for word in key.split('_'))
+        doc.add_heading(heading, level=1)
+        doc.add_paragraph(value)
+        # Add a line break between sections
+        doc.add_paragraph()
+    doc.save(filename)
